Venue Caching, Improved error pages, refactor
