+++ backend/app/ContentProviders/FourSquare.php
@@ -7,7 +7,8 @@
  */
 
 namespace App\ContentProviders;
-use App\DetailedVenue;
+
+use App\ContentProviders\FourSquare\VenueFactory;
 use App\Venue;
 use App\VenueCategory;
 use Illuminate\Support\Collection;
@@ -21,6 +22,19 @@ use Illuminate\Support\Facades\Log;
  */
 class FourSquare implements ContentProvider
 {
+    /**
+     * @var VenueFactory
+     */
+    protected $venueFactory;
+
+    /**
+     * FourSquare constructor.
+     * Initialize factories
+     */
+    public function __construct()
+    {
+        $this->venueFactory = new VenueFactory($this);
+    }
 
     /**
      * This will make a request to a certain API endpoint, and return the results.
@@ -62,27 +76,6 @@ class FourSquare implements ContentProvider
     }
 
     /**
-     * Turns a FourSquare object into a generic Venue Object
-     * @param $object
-     * @return Venue
-     */
-    public function createVenueFromAPIListing($object): Venue
-    {
-        $venue = new Venue();
-        $this->setBasicVenueItems($object, $venue);
-        return $venue;
-    }
-
-    public function createDetailedVenueFromAPIListing($object): Venue
-    {
-        $venue = new DetailedVenue();
-        $this->setBasicVenueItems($object, $venue);
-        $this->setDetailedVenueItems($object, $venue);
-        return $venue;
-    }
-
-    /**
-     * Obtain a number of collections, from a given long-latitude
      * @param float $longitude
      * @param float $latitude
      * @param int $radius
@@ -90,25 +83,16 @@ class FourSquare implements ContentProvider
      */
     public function getVenuesByLongLatitude(float $longitude, float $latitude, int $radius): array
     {
-        $endpoint = "venues/search";
-        $result = $this->get($endpoint, "GET", [
-            'll' => $latitude . "," . $longitude,
-            'intent' => "browse",
-            'radius' => $radius
-            ]);
+        return $this->venueFactory->getVenuesByLongLatitude($longitude, $latitude, $radius);
+    }
 
-        if($this->checkResponse($endpoint, $result) == false)
-            return [];
-
-        // For each venue we obtain, create a venue object
-        //dd($result);
-        $venues = [];
-        foreach ($result->response->venues as $venue) {
-            $venue = $this->createVenueFromAPIListing($venue);
-            array_push($venues,  $venue->getAsSimpleArray());
-        }
-
-        return $venues;
+    /**
+     * @param string $id
+     * @return Venue
+     */
+    public function getVenueById(string $id): Venue
+    {
+        return $this->venueFactory->getVenueById($id);
     }
 
     /**
@@ -140,62 +124,7 @@ class FourSquare implements ContentProvider
 
         // Bad Response
         Log::critical("FourSquare API call to endpoint: " . $endpoint . " failed. Information on line below. \r\n" . $request->meta->errorDetail);
+        abort($request->meta->code, $request->meta->errorDetail);
         return false;
-    }
-
-    /**
-     * @param $object
-     * @param $venue
-     */
-    private function setBasicVenueItems($object, Venue $venue): void
-    {
-        $venue->setId($object->id);
-        $venue->setName($object->name);
-        $venue->setAddressHumanReadable($object->location->formattedAddress);
-        $venue->setLatitude($object->location->lat);
-        $venue->setLongitude($object->location->lng);
-    }
-
-    /**
-     * Add categories to the actual venue
-     * @param $object
-     * @param DetailedVenue $venue
-     */
-    private function setDetailedVenueItems($object, DetailedVenue $venue)
-    {
-        $this->setDetailedVenueCategories($object->categories, $venue);
-        $venue->setURL($object->url);
-        $venue->setOpeningHours($object->hours->status);
-        if(isset($object->popular->status)) $venue->setPopularHours($object->popular->status);
-        if(isset($object->price)) $venue->setPriceClass($object->price->message);
-        $venue->setRating($object->rating);
-        $venue->setRatingColor($object->ratingColor);
-        $venue->setPeopleNow($object->hereNow->count);
-        $venue->setLikes($object->likes->count);
-    }
-
-    /**
-     * Obtain a DetailedVenue object by id
-     * @param string $id
-     * @return mixed
-     */
-    public function getVenueDetailsById(string $id)
-    {
-        $endpoint = 'venues/' . $id;
-        $result = $this->get($endpoint, "GET", []);
-
-        $detailedVenue = $this->createDetailedVenueFromAPIListing($result->response->venue);
-        return $detailedVenue;
-    }
-
-    private function setDetailedVenueCategories($categories, DetailedVenue $venue)
-    {
-        foreach($categories as $category) {
-            $newCategory = new VenueCategory();
-            $newCategory->setIcon($category->icon->prefix . "64" . $category->icon->suffix);
-            $newCategory->setName($category->name);
-            $newCategory->setPluralName($category->pluralName);
-            array_push($venue->categories, $newCategory);
-        }
     }
 }
